+++ src/main/webapp/app/tasks/services/default-task.service.ts
@@ -2,9 +2,9 @@ import { HttpClient } from '@angular/common/http';
 import { Inject, Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 
-import { BASE_URL } from '../app.tokens';
-import { Task } from './task';
-import { TaskService } from './task.service';
+import { BASE_URL } from '../../app.tokens';
+import { Task } from '../task';
+import { TaskService } from '../task.service';
 
 @Injectable()
 export class DefaultTaskService implements TaskService {
