+++ src/main/webapp/app/tasks/services/local-task.service.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
-import { LocalTaskService } from 'app/tasks/local-task.service';
+import { LocalTaskService } from 'app/tasks/services/local-task.service';
 import { Observable } from 'rxjs';
-import { Task } from './task';
+import { Task } from '../task';
 
 describe('LocalTaskService', () => {
   const id = 'de4f576e-d1b5-488a-8c77-63d4c8726909';
